requirements.txt
streamlit==1.37.1
scikit-learn==1.5.1
PyPDF2==3.0.1
python-docx==1.1.2